### 2015-03-17 #5 @karthik
SAMPLE ISSUE (testing)
* 1. What does this package do? (explain in 50 words or less)
Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
tempor incididunt ut labore
* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).
```
Package: rglobi
Title: R Interface to Global Biotic Interactions
Description: A programmatic interface to the web service methods
    provided by Global Biotic Interactions (GloBI). GloBI provides
    access to spatial-temporal species interaction records from
    sources all over the world. rglobi provides methods to search
    species interactions by location, interaction type, and
    taxonomic name. In addition, it supports Cypher, a graph query
    language, to allow for executing custom queries on the GloBI
    aggregate species interaction dataset.
Version: 0.2.6
Authors@R: c(person("Jorrit", "Poelen", role = c("aut", "cre"),
    email = "jhpoelen@xs4all.nl"),
    person("Stephen", "Gosnell", role = "aut",
    email = "stephen.gosnell@baruch.cuny.edu"),
    person("Sergey", "Slyusarev", role = "aut",
    email = "cph.lmy@gmail.com"))
    rjson (>= 0.2.13),
    RCurl
    knitr
```
* 3. URL for the package (the development repository, not a stylized html page): https://github.com/ropensci/rglobi
* 4. What data source(s) does it work with (if applicable)? None
* 5. Who is the target audience? Your doge
* 6. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [x] The repository has continuous integration with Travis and/or another service
* [x] The package contains a vignette
* [x] The package contains a complete readme with devtools install instructions
* [x] The package contains unit tests
* 7. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
* [x] Does the package have a CRAN accepted license?
* [ ] Are there any packages that your package depends on that are not on CRAN?
* 8. Please add explanations below for any exceptions to the above:

Nothing further to add. Thanks for considering!
