### 2015-03-17 #5 @sckott
looks good to me. I guess we have to add labels, or can the user add a label? if so, would be good to add a label e.g. https://github.com/ropensci/onboarding/labels/package
