### 2015-03-17 #5 @karthik
But we could do that too after an initial scan of the 50 word description
