### 2015-03-17 #5 @jennybc
Ask if it passes R CMD check in present state? That will catch/surface lots of things, such as missing function level documentation, right?
