### 2015-03-17 #5 @sckott
@jennybc  That makes sense to do - could say don't' submit until no warnings
