### 2015-03-17 #5 @karthik
@jennybc Good idea. Should we make this optional? Like _hey I want to get this package to rOpenSci but there is this one pesky warning that won't go away and I don't know how_

I've helped a couple of people that way. OR should we just help where the repo exists and only takes ones passing `CHECK` at this stage of the process.
