### 2015-03-17 #5 @jennybc
Proposal: ask submitter to reveal errors, warnings, and/or notes. Don't accept until no errors. At your discretion, help people work through warnings and notes, before or after acceptance.
