### 2015-03-18 #5 @karthik
@sckott @jennybc @richfitz  If you all think this looks ok and we've covered the major bases, I think it's time to make this the official route to submit a package to rOpenSci. As Scott might have mentioned, we already have a few packages in the queue that need to get pushed through.

I don't think we have to get everything perfectly ready, just as long as we keep iterating till this looks good (the submission process I imagine will continuously evolve, much like CRAN, although more friendly).
