### 2015-03-18 #5 @karthik
I suppose we should revserse the order of these two questions

Does the package have a CRAN accepted license?
 Do you intend for this package to go on CRAN?
