### 2015-03-19 #5 @sckott
should we redo @richfitz issue with the new format?  or just sorta move on with #1  as is
