### 2015-03-25 8f6aa63
added `accordingto` parameter to get_interactions_by_taxa
