### 2015-03-25 c87337d
lower case showField parameter to having consistent naming
