### 2015-03-25 5c49fd7
arguments are now passed to get_interactions_by_taxa
