### 2015-03-26 b81caaf
`otherkeys` argumant added to `get_interactions_by_taxa` function
This optional argument can be used for any key-value pair not listed in main function arguments in a form `otherkeys = list ("somekey" = c("value1", "value2", ...), ...`. So we can limit number of  `get_interactions_by_taxa` arguments without humpering its functionality.
