### 2015-03-26 637e6c2
documentation and tests updated, get_data_fields function added
