### 2015-03-26 e65edff
update Rd and NAMESPACE after running roxygenize(".")
