+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2 (4.1.0): do not edit by hand
 
 export(get_child_taxa)
+export(get_data_fields)
 export(get_interaction_areas)
 export(get_interaction_matrix)
 export(get_interaction_table)
