### 2015-03-26 9082909
added test for interactionFields / get_data_fields; added autogenerated .Rd file; thanks to @jemmybutton for adding the docs
