+++ man/get_data_fields.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2 (4.1.0): do not edit by hand
+% Please edit documentation in R/rglobi.R
+\name{get_data_fields}
+\alias{get_data_fields}
+\title{List data fields identified in GloBI database}
+\usage{
+get_data_fields()
+}
+\value{
+Returns data frame of supported data fields
+}
+\description{
+Returns data frame with supported data fields
+}
+\examples{
+\dontrun{
+get_data_fields()
+}
+}
+\keyword{database}
+
