### 2015-03-30 #5 @richfitz
I vote for starting again so we can point people at a relatively clean process as an example.
